db recyclerView

+++ app/src/main/java/com/example/expensesmanager/db/Expense.kt
@@ -8,7 +8,7 @@ import java.sql.Timestamp
 data class Expense(
     @PrimaryKey
     val id: Int,
-    val amount: Double,
+    val amount: Float,
     val spent: Boolean,
     val time: Float
 )
